feature: use ha built in media browser
It has become an increasingly hard task to try and maintain a self-written media-browser and try to match what Home Assistant's built-in media browser can do. This area has been the one with the most bugs and issues reported. Because of this the decision is to only show favorites in media section, and for other media the user can use HA's built-in media browser dialog.